progress: 47.28% of code matched
Source-commit: b998dc7ccab642946213c3acb506145ce8043c61

+++ src/game/game_fn_801028A8.c
@@ -19,7 +19,7 @@ typedef struct BitReader {
 __declspec(section ".sdata") extern s32 lbl_8064CC68;
 __declspec(section ".sdata") extern s32 lbl_8064CC6C;
 
-static inline int read_bit(BitReader* reader)
+static inline s16 read_bit(BitReader* reader)
 {
     u8 bit;
 
@@ -34,8 +34,8 @@ static inline int read_bit(BitReader* reader)
 
 /* NonMatching: behavior-complete recursive decode-tree loader. Explicit inline
  * suppresses the unused out-of-line helper, making the unit's .text size exact.
- * The function differs only in MWCC's 0x40-byte frame and epilogue scheduling;
- * retail uses a 0x30-byte frame. */
+ * A 16-bit helper result selects retail's 0x30-byte frame. The sole remaining
+ * difference is MWCC's epilogue scheduling of mtlr after the final restores. */
 s16 fn_801028A8(DecodeTree* tree, BitReader* reader)
 {
     s16 value;
